# unicorn-api/docs/swagger.yaml
basePath: /
definitions:
...
  models.ComputeContainerInfo:
    properties:
      id:
        type: string
      image:
        type: string
      ports:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
    type: object
  models.ComputeCreateRequest:
    properties:
      expose_port:
        type: string
      image:
        type: string
      ports:
        additionalProperties:
          type: string
        type: object
      preset:
        allOf:
        - $ref: '#/definitions/models.ComputePreset'
        enum:
        - micro
        - small
    required:
    - expose_port
    - image
    - ports
    - preset
    type: object
  models.ComputePreset:
    enum:
    - micro
    - small
    type: string
    x-enum-varnames:
    - PresetMicro
    - PresetSmall
...
host: localhost:8080
info:
  contact:
    email: support@unicorn-api.com
    name: Unicorn API Support
    url: https://github.com/your-org/unicorn-api
  description: A comprehensive RESTful API for Unicorn services providing Identity
    and Access Management (IAM) functionality. This API supports user authentication,
    role-based access control, organization management, and JWT token handling.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Unicorn API
  version: "1.0"
paths:
...
  /api/v1/compute/create:
    post:
      consumes:
      - application/json
      description: Create a new compute container with the specified image and configuration
      parameters:
      - description: Compute container creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ComputeCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Container ID
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a compute container
      tags:
      - Compute
  /api/v1/compute/list:
    get:
      consumes:
      - application/json
      description: List all compute containers owned by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of containers
          schema:
            items:
              $ref: '#/definitions/models.ComputeContainerInfo'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List compute containers
      tags:
      - Compute
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"